+++ Changelog.md
@@ -2,6 +2,11 @@
 
 _Also check out the [roadmap](Roadmap.md) with features I’d like to add in the future._
 
+## 12 April 2022
+
+-   Tweaked API styling
+-   Added Obsidian plugin to API uses list
+
 ## 10 April 2022 v2.3
 
 -   [Rickies API](https://rickies.co/api) available to fetch the current chairmen in JSON format